Add explicit CRD configuration for MetalLB (#386)
Fixes #385 
Fixes #387 

The old MetalLB version detection code relied on permissions that the CCM doesn't have in order.  Instead, this adds support for an explicit `crdConfiguration` flag as a URL parameter in the `loadBalancer` configuration string.  Initially, the `crdConfiguration` flag defaults to `false` for backwards compatibility; we will switch the default in a future major release.

In addition, this fixes the MetalLB configuration code so that the documented namespace-only configuration (`metallb:///<namespace>`), which is essential for MetalLB 0.13.2+, is supported