fix: patch service object instead of updating
When deploying a new LoadBalancer service, we would see a "Failed
to update..." error message in the logs, and 2 load balancers would
be created on the Equinix side, with only 1 of them reflected in
the service annotations.

This changes our `Update` call to update the k8s service object with
a `Patch` call; by doing this we focus our k8s API calls on changing
only the properties that our provider has modified, which reduces the
chance that something else has concurrently made conflicting changes.