refactor: defer implementation of GetLoadBalancer to EMLB
To avoid adding knowledge of specific load balancer implementers in
`loadbalancers.go`, this moves the `GetLoadBalancer` logic for EMLB
into `emlb.go`.

In the future we can refactor the pre-existing `GetLoadBalancer`
code in order to push that down to `metallb`, `kubevip`, `empty`,
etc., keeping `loadbalancers.go` focused on the work of deciding
which implementer to use and how to translate from the cloud-provider
interface to our internal implementer interface.