fix: actually delete old resources when updating a service
After inspecting API responses in the Equinix Metal console, I saw
that origin pool data is not nested under the corresponding port.

Instead, `ports` is an array of ports, and `pools` is an array of
arrays of pools.  The top-level index of `pools` corresponds to
the top-level index of `ports`, so `pools[i]` is an array of all
of the origin pools assigned to `ports[i]`.

The old code was looking for origin data in the wrong place, so it
was failing to delete origins that are no longer needed.  Now it
is looking in the right place and successfully deleting old origins.