fix: return a valid LoadBalancerStatus for BGP-based load balancers
While introducing Equinix Metal Load Balancer support in CPEM, we updated
the `EnsureLoadBalancer` function to return the result of `GetLoadBalancer`
so that the code to return a LoadBalancerStatus for existing load balancers
is in one place.

As a result of that consolidation, we uncovered an inconsistency in the two
calculations of LoadBalancerStatus: the one in EnsureLoadBalancer created an
Ingress with the load balancer IP, but the one in GetLoadBalancer created an
Ingress with the load balancer CIDR, which is not valid.

This updates the GetLoadBalancer implementation to match the valid code that
was previously in EnsureLoadBalancer.