enable equinix-sdk-go debugging when PACKNGO_DEBUG env var is set
The old packngo SDK had built-in support for enabling debug logging if the
`PACKNGO_DEBUG` environment variable was set to any value.  The generated
equinix-sdk-go SDK requires explicitly enabling debug logging in consumer
code, so when we switched from `packngo` to `equinix-sdk-go` we lost support
for the `PACKNGO_DEBUG` environment variable.  This updates CPEM to enable
debug logging for `equinix-sdk-go` if the old `PACKNGO_DEBUG` environment
variable is set.